Golem pathfinding bug
GetRect de Skeleton ara torna 2 floats en canvi de int!

+++ cSkeleton.h
@@ -20,7 +20,7 @@ public:
 	void GoToCell(int *map, int destcx, int destcy,int type); //type= 0 - normal, 1 - attack
 	void GoToPlayer(int *map, int destcx, int destcy);
 	void Move();
-	void GetRect(RECT *rc,int *posx,int *posy,cScene *Scene);
+	void GetRect(RECT *rc,float *posx,float *posy,cScene *Scene);
 	void GetRectRadar(RECT *rc,int *posx,int *posy);
 
 	void SetPosition(int posx,int posy);
@@ -42,7 +42,7 @@ private:
 	int sightRadius;		//Radius of sight
 	int x,y;				//Position in total map
 	int cx,cy;				//Cell position in total map in cartesian
-	int ix, iy;				//Position in total map in isometric
+	float ix, iy;			//Position in total map in isometric
 	int sprite_height;		//Offset to aply on the vertical axis in case the sprite is larger than 1 tile
 
 	bool active;
